// packages/editor/editor-core/src/plugins/history-analytics/pm-plugins/prosemirror-history-types.ts
import { StepMap, Step } from 'prosemirror-transform';
import { SelectionBookmark } from 'prosemirror-state';

export const pmHistoryPluginKey = 'history$';
export const DEPTH_OVERFLOW = 20;

export interface PmHistoryItem {
  map: StepMap;
  mirrorOffset?: number;
  selection?: SelectionBookmark;
  step: Step;
}

export interface PmHistoryLeaf {
  values: PmHistoryItem[];
  length: number;
  forEach: (
    handler: (item: PmHistoryItem) => boolean | void,
    from?: number,
    to?: number,
  ) => void;
}

export interface PmHistoryBranch {
  items: PmHistoryLeaf;
  eventCount: number;
}

export interface PmHistoryPluginState {
  done: PmHistoryBranch;
  undone: PmHistoryBranch;
  prevMap?: StepMap;
  prevTime: number;
}

export interface PmHistoryPluginConfig {
  depth: number;
  newGroupDelay: number;
}